requirements.txt
Appium-Python-Client==4.0.0
selenium==4.22.0
PyYAML==6.0.2
APScheduler==3.10.4
requests==2.32.3
jinja2==3.1.4
